use Cookies

+++ blackjack/core/src/main/scala/de/htwg/core/entities/Bet.scala
@@ -3,13 +3,13 @@ package de.htwg.core.entities
 /**
   * Created by Michael Meister on 09.01.2016.
   */
-class Bet(private var amount: Double) {
+class Bet(private var amount: Int) {
 
-  def +(amount: Double): Unit = {
+  def +(amount: Int): Unit = {
     this.amount += amount
   }
 
-  def getAmount(): Double = {
+  def getAmount(): Int = {
     amount
   }
 }
